requirements.txt
Flask==2.2.2
opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-exporter-jaeger==1.17.0
